+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 10.0.0-dev.0
+
+- Add support for ObjC Blocks that can be invoked from any thread, using
+  NativeCallable.listener.
+- Bump min SDK version to 3.2.0-114.0.dev.
+
 # 9.0.1
 
 - Fix doc comment missing on struct/union array fields.